Moved all imposter-related classes into new Models.Imposters namespace
