+++ MbDotNet.Tests/MountebankRequestProxyTests.cs
@@ -2,6 +2,7 @@
 using MbDotNet.Enums;
 using MbDotNet.Exceptions;
 using MbDotNet.Models;
+using MbDotNet.Models.Imposters;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using RestSharp;
